+++ ChangeLog.md
@@ -8,6 +8,8 @@
   ([#71](https://github.com/davep/textual-fspicker/pull/71))
 - Added Python 3.14 as a tested/supported Python version.
   ([#72](https://github.com/davep/textual-fspicker/pull/72))
+- Added `textual_fspicker.Icons` so that the user of the library can set
+  their own icons that are displayed next to directory entries.
 
 ## v0.5.0